Master Script

Calls all other functions for the ContDataQC library.

ContDataQC(
  fun.myData.Operation,
  fun.myData.SiteID,
  fun.myData.Type,
  fun.myData.DateRange.Start,
  fun.myData.DateRange.End,
  fun.myDir.import = getwd(),
  fun.myDir.export = getwd(),
  fun.myConfig = "",
  fun.myFile = "",
  fun.myReport.format = "",
  fun.myReport.Dir = "",
  fun.CreateReport = TRUE
)

Arguments

fun.myData.Operation
Operation to be performed; c("GetGageData","QCRaw", "Aggregate", "SummaryStats")

fun.myData.SiteID
Station/SiteID.

fun.myData.Type
data type; c("Air","Water","AW","Gage","AWG","AG","WG")

fun.myData.DateRange.Start
Start date for requested data. Format = YYYY-MM-DD.

fun.myData.DateRange.End
End date for requested data. Format = YYYY-MM-DD.

fun.myDir.import
Directory for import data. Default is current working directory.

fun.myDir.export
Directory for export data. Default is current working directory.

fun.myConfig
Configuration file to use for this data analysis. The default is always loaded first so only "new" values need to be included. This is the easiest way to control time zones.

fun.myFile
Single file (or vector of files) to perform functions. SiteID, Type, and Date Range not used when file name(s) provided.

fun.myReport.format
Report format (docx or html). Default is specified in config.R (docx). Can be customized in config.R; ContData.env$myReport.Format.

fun.myReport.Dir
Report (rmd) template folder. Default is the package rmd folder. Can be customized in config.R; ContData.env$myReport.Dir.

fun.CreateReport
Boolean parameter to create reports or not. Default = TRUE.

Value

Returns a csv into the specified export directory with additional columns for calculated statistics.

Details

Below are the default data directories assumed to exist in the working directory. These can be created with code in the example.

./Data0_Original/ = Unmodified data logger files.

./Data1_RAW/ = Data logger files modified for use with library. Modifications for extra rows and file and column names.

./Data2_QC/ = Repository for library output for QCed files.

./Data3_Aggregated/ = Repository for library output for aggregated (or split) files.

./Data4_Stats/ = Repository for library output for statistical summary files.

It is possible to call the Aggregate file portion of the script to meld together files from multiple sites (e.g., all sites in a watershed or from different depths on a lake). The file will be named by the first file named with "Append_x" where "x" is the number of files that were aggregated. The purpose is to allow users to analyze the data in these files from a single file.

Examples