Export data for IHA package

Creates a date frame (and file export) from Continuous Data in the format used by the IHA package.

Export.IHA(
  fun.myFile,
  fun.myDir.import = getwd(),
  fun.myDir.export = getwd(),
  fun.myDateRange.Start = NA,
  fun.myDateRange.End = NA,
  fun.myCol.DateTime = "Date.Time",
  fun.myCol.Parameter = "Discharge.ft3.s"
)

Arguments

fun.myFile
Filename (no directory) of data file. Must be CSV file. Includes Date/Time and Discharge fields.

fun.myDir.import
Directory for import data. Default is current working directory.

fun.myDir.export
Directory for export data. Default is current working directory.

fun.myDateRange.Start
Start date for IHA analysis. File will be filtered on this date. Default is NA (no filtering).

fun.myDateRange.End
End date for IHA analysis. File will be filtered on this date. Default is NA (no filtering).

fun.myCol.DateTime
Column name in myFile with Date/Time. Assumes format of %Y-%m-%d %H:%M:%S. Default = "Date.Time".

fun.myCol.Parameter
Column name in myFile for measurements. Default = "Discharge.ft3.s".

fun.myDate.Format
Format of benchmark date. This should be the same format of the date in the data file. Default is %Y-%m-%d (e.g., 2017-12-31).

Value

Returns a data frame with daily mean values by date (in the specified range). Also, a csv file is saved to the specified directory with the prefix "IHA" and the date range before the file extension.

Details

The function assumes the provided data is sampled more often then daily. The daily means are calculated internally.

The IHA package is not included in the ContDataQC package. But an example is provided.

To run the example IHA calculations you will need the IHA package (from GitHub) and for the example export the XLConnect packages Install commands in the example.

Examples

# 1.  Get Gage Data
#
# 1.A. Use ContDataQC and Save (~1min for download)
myData.Operation    <- "GetGageData" #Selection.Operation[1]
myData.SiteID       <- "01187300" # Hubbard River near West Hartland, CT
myData.Type         <- "Gage"
myData.DateRange.Start  <- "2015-01-01"
myData.DateRange.End    <- "2016-12-31"
myDir.import <- getwd()
myDir.export <- getwd()
ContDataQC(myData.Operation, myData.SiteID, myData.Type
           , myData.DateRange.Start, myData.DateRange.End
           , myDir.import, myDir.export)

#> [1] "Total items to process = 1"
#> 
#> [1] "Getting available data; 01187300."
#> 
#>     agency_cd  site_no                           station_nm site_tp_cd
#> 530      USGS 01187300 HUBBARD RIVER NEAR WEST HARTLAND, CT         ST
#> 531      USGS 01187300 HUBBARD RIVER NEAR WEST HARTLAND, CT         ST
#> 532      USGS 01187300 HUBBARD RIVER NEAR WEST HARTLAND, CT         ST
#>     dec_lat_va dec_long_va coord_acy_cd dec_coord_datum_cd  alt_va alt_acy_va
#> 530    42.0375   -72.93933            H              NAD83  594.57       0.01
#> 531    42.0375   -72.93933            H              NAD83  594.57       0.01
#> 532    42.0375   -72.93933            H              NAD83  594.57       0.01
#>     alt_datum_cd   huc_cd data_type_cd parm_cd stat_cd ts_id loc_web_ds
#> 530       NGVD29 01080207           uv   00010    <NA> 66869         NA
#> 531       NGVD29 01080207           uv   00060    <NA> 66866         NA
#> 532       NGVD29 01080207           uv   00065    <NA> 66867         NA
#>     medium_grp_cd parm_grp_cd   srs_id access_cd begin_date   end_date count_nu
#> 530           wat        <NA>  1645597         0 2015-11-04 2020-01-27     1545
#> 531           wat        <NA>  1645423         0 1990-10-01 2020-01-27    10710
#> 532           wat        <NA> 17164583         0 2007-10-01 2020-01-27     4501
#> 
#> 
#> [1] "Processing item 1 of 1, COMPLETE, 01187300."
#> 
#> [1] "Task COMPLETE; 0.63 min."

#
# 1.B. Use saved data
myData.SiteID   <- "01187300"
myFile          <- "01187300_Gage_20150101_20161231.csv"
myCol.DateTime  <- "Date.Time"
myCol.Discharge <- "Discharge.ft3.s"
#
# 2. Prep Data
myData.IHA <- Export.IHA(myFile
                         , fun.myCol.DateTime = myCol.DateTime
                         , fun.myCol.Parameter = myCol.Discharge
                         )

#> Error in read.csv(myFile, stringsAsFactors = FALSE): object 'myFile' not found

#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# 3. Run IHA
# Example using returned DF with IHA
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# User info
SiteID <- myData.SiteID
Notes.User <- Sys.getenv("USERNAME")
#~~~~~
# Library (install if needed)
# devtools::install_github("jasonelaw/IHA")
# install.packages("XLConnect")
# Library (load)
require(IHA)

#> Loading required package: IHA

#> Loading required package: zoo

#> 
#> Attaching package: 'zoo'

#> The following objects are masked from 'package:base':
#> 
#>     as.Date, as.Date.numeric

require(XLConnect)

#> Loading required package: XLConnect

#> Loading required package: XLConnectJars

#> Error: package or namespace load failed for 'XLConnectJars':
#>  .onLoad failed in loadNamespace() for 'rJava', details:
#>   call: fun(libname, pkgname)
#>   error: JAVA_HOME cannot be determined from the Registry

#~~~~~
# IHA
myYr <- "calendar" # "water" or "calendar"
# IHA Metrics
## IHA parameters group 1; Magnitude of monthly water conditions
Analysis.Group.1 <- group1(myData.IHA, year=myYr)

#> Error in is.zoo(x): object 'myData.IHA' not found

## IHA parameters group 2: Magnitude of monthly water condition and include 12 parameters
Analysis.Group.2 <- group2(myData.IHA, year=myYr)

#> Error in is.zoo(x): object 'myData.IHA' not found

Analysis.Group.3 <- group3(myData.IHA, year=myYr)

#> Error in is.zoo(x): object 'myData.IHA' not found

## IHA parameters group 4; Frequency and duration of high and low pulses
# defaults to 25th and 75th percentiles
Analysis.Group.4 <- group4(myData.IHA, year=myYr)

#> Error in is.zoo(x): object 'myData.IHA' not found

## IHA parameters group 5; Rate and frequency of water condition changes
Analysis.Group.5 <- group5(myData.IHA, year=myYr)

#> Error in is.zoo(x): object 'myData.IHA' not found

#~~~~~
# Save Results to Excel (each group on its own worksheet)
Group.Desc <- c("Magnitude of monthly water conditions"
                ,"Magnitude of monthly water condition and include 12 parameters"
                ,"Timing of annual extreme water conditions"
                ,"Frequency and duration of high and low pulses"
                ,"Rate and frequency of water condition changes")
df.Groups <- as.data.frame(cbind(paste0("Group",1:5),Group.Desc))
#
myDate <- format(Sys.Date(),"%Y%m%d")
myTime <- format(Sys.time(),"%H%M%S")
# Notes section (add min/max dates)
Notes.Names <- c("Dataset (SiteID)","IHA.Year","Analysis.Date (YYYYMMDD)"
                 ,"Analysis.Time (HHMMSS)","Analysis.User")
Notes.Data <- c(SiteID, myYr, myDate, myTime, Notes.User)
df.Notes <- as.data.frame(cbind(Notes.Names,Notes.Data))
Notes.Summary <- summary(myData.IHA)

#> Error in summary(myData.IHA): object 'myData.IHA' not found

# Open/Create file
myFile.XLSX <- paste("IHA", SiteID, myYr, myDate, myTime, "xlsx", sep=".")
wb <- loadWorkbook(myFile.XLSX, create = TRUE) # load workbook, create if not existing

#> Error in loadWorkbook(myFile.XLSX, create = TRUE): could not find function "loadWorkbook"

# create sheets
createSheet(wb, name = "NOTES")

#> Error in createSheet(wb, name = "NOTES"): could not find function "createSheet"

createSheet(wb, name = "Group1")

#> Error in createSheet(wb, name = "Group1"): could not find function "createSheet"

createSheet(wb, name = "Group2")

#> Error in createSheet(wb, name = "Group2"): could not find function "createSheet"

createSheet(wb, name = "Group3")

#> Error in createSheet(wb, name = "Group3"): could not find function "createSheet"

createSheet(wb, name = "Group4")

#> Error in createSheet(wb, name = "Group4"): could not find function "createSheet"

createSheet(wb, name = "Group5")

#> Error in createSheet(wb, name = "Group5"): could not find function "createSheet"

# write to worksheet
writeWorksheet(wb, df.Notes, sheet = "NOTES", startRow=1)

#> Error in writeWorksheet(wb, df.Notes, sheet = "NOTES", startRow = 1): could not find function "writeWorksheet"

writeWorksheet(wb, Notes.Summary, sheet = "NOTES", startRow=10)

#> Error in writeWorksheet(wb, Notes.Summary, sheet = "NOTES", startRow = 10): could not find function "writeWorksheet"

writeWorksheet(wb, df.Groups, sheet="NOTES", startRow=25)

#> Error in writeWorksheet(wb, df.Groups, sheet = "NOTES", startRow = 25): could not find function "writeWorksheet"

writeWorksheet(wb, Analysis.Group.1, sheet = "Group1")

#> Error in writeWorksheet(wb, Analysis.Group.1, sheet = "Group1"): could not find function "writeWorksheet"

writeWorksheet(wb, Analysis.Group.2, sheet = "Group2")

#> Error in writeWorksheet(wb, Analysis.Group.2, sheet = "Group2"): could not find function "writeWorksheet"

writeWorksheet(wb, Analysis.Group.3, sheet = "Group3")

#> Error in writeWorksheet(wb, Analysis.Group.3, sheet = "Group3"): could not find function "writeWorksheet"

writeWorksheet(wb, Analysis.Group.4, sheet = "Group4")

#> Error in writeWorksheet(wb, Analysis.Group.4, sheet = "Group4"): could not find function "writeWorksheet"

writeWorksheet(wb, Analysis.Group.5, sheet = "Group5")

#> Error in writeWorksheet(wb, Analysis.Group.5, sheet = "Group5"): could not find function "writeWorksheet"

# save workbook
saveWorkbook(wb, myFile.XLSX)

#> Error in saveWorkbook(wb, myFile.XLSX): could not find function "saveWorkbook"