Quality Control, File

Subfunction for performing QC. Needs to be called from ContDataQC(). Requires zoo(). The 'file' version works on a single (or vector) of files.

fun.QC.File(
  fun.myFile,
  fun.myDir.import = getwd(),
  fun.myDir.export = getwd(),
  fun.myReport.format,
  fun.myReport.Dir,
  fun.CreateReport = TRUE
)

Arguments

fun.myFile
Single (or vector of) file name(s).

fun.myDir.import
Directory for import data. Default is current working directory.

fun.myDir.export
Directory for export data. Default is current working directory.

fun.myReport.format
Report format (docx or html). Default is specified in config.R (docx).Can be customized in config.R; ContData.env$myReport.Format.

fun.myReport.Dir
Report (rmd) template folder. Default is the package rmd folder. Can be customized in config.R; ContData.env$myReport.Dir.

fun.CreateReport
Boolean parameter to create reports or not. Default = TRUE.

Value

Returns a csv file to specified directory with QC flags.

Examples

#Not intended to be accessed indepedant of function ContDataQC().
myFile <- "test2_AW_20130426_20130725.csv"
#myFile <- c("test2_AW_20130426_20130725.csv", "test2_AW_20130725_20131015.csv"
#             , "test2_AW_20140901_20140930.csv")
myDir.import <- file.path(getwd(),"Data1_RAW")
myDir.export <- file.path(getwd(),"Data2_QC")
fun.QC.File(myFile, myDir.import, myDir.export)

#> [1] "Total files to process = 1"
#> [1] "ERROR; no such file exits.  Cannot QC the data."
#> [1] "PATH = /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp95vHKB/file8d449c56c5/reference/Data1_RAW"
#> [1] "FILE = test2_AW_20130426_20130725.csv"

#> Error in fun.QC.File(myFile, myDir.import, myDir.export): Bad file.