Statistical Summary, File

Subfunction for generating statistical summaries. Needs to be called from ContDataQC(). Requires doBy() and survival() [required by doBy] Calculates statistics on input data and saves to a new csv. Plots are saved as PDF. The 'file' version works on a single (or vector) of files.

fun.Stats.File(
  fun.myFile,
  fun.myDir.import = getwd(),
  fun.myDir.export = getwd(),
  fun.myReport.format,
  fun.myReport.Dir
)

Arguments

fun.myFile

Single (or vector of) file name(s).

fun.myDir.import

Directory for import data. Default is current working directory.

fun.myDir.export

Directory for export data. Default is current working directory.

fun.myReport.format

Report format (docx or html). Default is specified in config.R (docx). Can be customized in config.R; ContData.env$myReport.Format.

fun.myReport.Dir

Report (rmd) template folder. Default is the package rmd folder. Can be customized in config.R; ContData.env$myReport.Dir.

Value

Returns a csv into the specified export directory with additional columns for calculated statistics. And a PDF of summary plots by different time periods.

Examples

myFile <- "QC_test2_Aw_20130426_20130725.csv"
#myFile <- c("QC_test2_Aw_20130426_20130725.csv"
#           , "QC_test2_Aw_20130725_20131015.csv", "QC_test2_Aw_20140901_20140930.csv")
myDir.import <- file.path(getwd(),"Data2_QC")
myDir.export <- file.path(getwd(),"Data4_Stats")
fun.Stats.File(myFile, myDir.import, myDir.export)
#> [1] "Total files to process = 1"
#> [1] "ERROR; no such file exits.  Cannot generate summary statistics."
#> [1] "PATH = /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpMk2p0Z/file6f41fd8f08e/reference/Data2_QC"
#> [1] "FILE = QC_test2_Aw_20130426_20130725.csv"
#> Error in fun.Stats.File(myFile, myDir.import, myDir.export): Bad file.