Richards-Baker (Flashiness) Index [RBI] calculator

Input is a vector of mean daily flows. Output is a value that is the RB Flashiness Index

RBIcalc(Q)

Arguments

Q

Vector of mean daily stream flows.

Value

Returns a value that represents the RBI.

Details

The index is valid over the time period of the data. For example, if the vector contains a month's or a year's data the index will represent that time period. The index will be in the same units as the input data. The function assumes all days are represented (insert NA for missing values).

Baker, D.B., R.P. Richards, T.T. Loftus, and J.W. Kramer. 2004.

A New Flashiness Index: Characteristics and Applications to Midwestern Rivers and Streams. April 2004. Journal of the American Water Resources Association (JAWRA). Pages 503:522.

Examples

# Get Gage Data via the dataRetrieval package from USGS 01187300 2013
data.gage <- dataRetrieval::readNWISdv("03238500"
                                       , "00060"
                                       , "1974-10-01"
                                       , "1975-09-30")
head(data.gage)
#>   agency_cd  site_no       Date X_00060_00003 X_00060_00003_cd
#> 1      USGS 03238500 1974-10-01            38                A
#> 2      USGS 03238500 1974-10-02            32                A
#> 3      USGS 03238500 1974-10-03            26                A
#> 4      USGS 03238500 1974-10-04            20                A
#> 5      USGS 03238500 1974-10-05            23                A
#> 6      USGS 03238500 1974-10-06            23                A
# flow data
data.Q <- data.gage[,4]
# remove zeros
data.Q[data.Q==0] <- NA
RBIcalc(data.Q)
#> [1] 0.9833356

# QC with document Baker et al., 2004
# Table 1, Whiteoak Creek near Georgetown, Ohio (03238500)
# Figure 8, upward pointing triangle for 1975 water year value close to
# "1.0".  Using the data downloaded in journal example calculated RBI values
# in Excel and R match at 0.9833356.