Download USGS Gage Data

This a wrapper function for the package dataRetrieval to get USGS data. Daily means are the default data type.

fun.GageData(
  fun.myData.SiteID,
  fun.myData.Type = "Gage",
  fun.myData.DateRange.Start,
  fun.myData.DateRange.End,
  fun.myDir.export = getwd(),
  fun.myTZ = ContData.env$myTZ
)

Arguments

fun.myData.SiteID
Station/SiteID (can be a vector).

fun.myData.Type
data type is "Gage".

fun.myData.DateRange.Start
Start date for requested data. Format = YYYY-MM-DD.

fun.myData.DateRange.End
End date for requested data. Format = YYYY-MM-DD.

fun.myDir.export
Directory for export data. Default is current working directory.

fun.myTZ
Timezone for requested gage. Default is in env.UserDefinedValues.R. Can also be set with Sys.timezone().

Value

Returns a csv file to specified directory with the requested daily mean data. During the data retrieval a summary is output to the console.

Details

Gage height is converted to water level for compatability with HOBOware output.

Examples

#Not intended to be accessed independant of function ContDataQC().
fun.GageData("01187300","Gage","2013-01-01","2013-12-31",getwd(),"","")

#> Error in fun.GageData("01187300", "Gage", "2013-01-01", "2013-12-31",     getwd(), "", ""): unused argument ("")

# with optional variable left blank
fun.GageData("01187300","Gage","2013-01-01","2013-12-31")

#> [1] "Total items to process = 1"
#> 
#> [1] "Getting available data; 01187300."
#> 
#>     agency_cd  site_no                           station_nm site_tp_cd
#> 530      USGS 01187300 HUBBARD RIVER NEAR WEST HARTLAND, CT         ST
#> 531      USGS 01187300 HUBBARD RIVER NEAR WEST HARTLAND, CT         ST
#> 532      USGS 01187300 HUBBARD RIVER NEAR WEST HARTLAND, CT         ST
#>     dec_lat_va dec_long_va coord_acy_cd dec_coord_datum_cd  alt_va alt_acy_va
#> 530    42.0375   -72.93933            H              NAD83  594.57       0.01
#> 531    42.0375   -72.93933            H              NAD83  594.57       0.01
#> 532    42.0375   -72.93933            H              NAD83  594.57       0.01
#>     alt_datum_cd   huc_cd data_type_cd parm_cd stat_cd ts_id loc_web_ds
#> 530       NGVD29 01080207           uv   00010    <NA> 66869         NA
#> 531       NGVD29 01080207           uv   00060    <NA> 66866         NA
#> 532       NGVD29 01080207           uv   00065    <NA> 66867         NA
#>     medium_grp_cd parm_grp_cd   srs_id access_cd begin_date   end_date count_nu
#> 530           wat        <NA>  1645597         0 2015-11-04 2021-01-08     1892
#> 531           wat        <NA>  1645423         0 1990-10-01 2021-01-08    11057
#> 532           wat        <NA> 17164583         0 2007-10-01 2021-01-08     4848
#> 
#> 
#> [1] "Processing item 1 of 1, COMPLETE, 01187300."
#> 
#> [1] "Task COMPLETE; 0.09 min."